+++ src/platforms/pc/platform.c
@@ -82,7 +82,7 @@ static ret_t date_time_get_now_impl(date_time_t* dt) {
 
 #endif
 
-uint32_t get_time_ms() {
+uint64_t get_time_ms() {
   struct timeval tv;
   gettimeofday(&tv, NULL);
