+++ src/platforms/raw/sys_tick.c
@@ -21,19 +21,19 @@
 
 #include "tkc/types_def.h"
 
-static volatile uint32_t g_sys_tick;
+static volatile uint64_t g_sys_tick;
 
 void SysTick_Handler(void) {
   g_sys_tick++;
 }
 
-uint32_t get_time_ms() {
+uint64_t get_time_ms() {
   return g_sys_tick;
 }
 
 void sleep_ms(uint32_t ms) {
   uint32_t count = 0;
-  uint32_t start = get_time_ms();
+  uint64_t start = get_time_ms();
 
   while (get_time_ms() < (start + ms)) {
     count++;
